--- api-test/user.http
@@ -1,3 +1,0 @@
-
-### 测试
-GET http://antio2.cn
