+++ sfx-analyze/sql/init&data.sql
@@ -29,107 +29,4 @@ CREATE TABLE `watchlist`  (
 -- ----------------------------
 -- Records of watchlist
 -- ----------------------------
-INSERT INTO `watchlist` VALUES (0, 0x303031);
-INSERT INTO `watchlist` VALUES (0, 0x3030303032);
-INSERT INTO `watchlist` VALUES (0, 0x3030303033);
-INSERT INTO `watchlist` VALUES (0, 0x3030303034);
-INSERT INTO `watchlist` VALUES (0, 0x3030303035);
-INSERT INTO `watchlist` VALUES (0, 0x3030303036);
-INSERT INTO `watchlist` VALUES (0, 0x3030303037);
-INSERT INTO `watchlist` VALUES (0, 0x3030303038);
-INSERT INTO `watchlist` VALUES (0, 0x3030303039);
-INSERT INTO `watchlist` VALUES (0, 0x303030303130);
-INSERT INTO `watchlist` VALUES (0, 0x303030303131);
-INSERT INTO `watchlist` VALUES (0, 0x303030303132);
-INSERT INTO `watchlist` VALUES (0, 0x303030303133);
-INSERT INTO `watchlist` VALUES (0, 0x303030303134);
-INSERT INTO `watchlist` VALUES (0, 0x303030303135);
-INSERT INTO `watchlist` VALUES (0, 0x303030303136);
-INSERT INTO `watchlist` VALUES (0, 0x303030303137);
-INSERT INTO `watchlist` VALUES (0, 0x303030303138);
-INSERT INTO `watchlist` VALUES (0, 0x303030303139);
-INSERT INTO `watchlist` VALUES (0, 0x303030303230);
-INSERT INTO `watchlist` VALUES (0, 0x303030303231);
-INSERT INTO `watchlist` VALUES (0, 0x303030303232);
-INSERT INTO `watchlist` VALUES (0, 0x303030303233);
-INSERT INTO `watchlist` VALUES (0, 0x303030303234);
-INSERT INTO `watchlist` VALUES (0, 0x303030303235);
-INSERT INTO `watchlist` VALUES (0, 0x303030303236);
-INSERT INTO `watchlist` VALUES (0, 0x303030303237);
-INSERT INTO `watchlist` VALUES (0, 0x303030303238);
-INSERT INTO `watchlist` VALUES (0, 0x303030303239);
-INSERT INTO `watchlist` VALUES (0, 0x303030303330);
-INSERT INTO `watchlist` VALUES (0, 0x303030303331);
-INSERT INTO `watchlist` VALUES (0, 0x303030303332);
-INSERT INTO `watchlist` VALUES (0, 0x303030303333);
-INSERT INTO `watchlist` VALUES (0, 0x303030303334);
-INSERT INTO `watchlist` VALUES (0, 0x303030303335);
-INSERT INTO `watchlist` VALUES (0, 0x303030303336);
-INSERT INTO `watchlist` VALUES (0, 0x303030303337);
-INSERT INTO `watchlist` VALUES (0, 0x303030303338);
-INSERT INTO `watchlist` VALUES (0, 0x303030303339);
-INSERT INTO `watchlist` VALUES (0, 0x303030303430);
-INSERT INTO `watchlist` VALUES (0, 0x303030303431);
-INSERT INTO `watchlist` VALUES (0, 0x303030303432);
-INSERT INTO `watchlist` VALUES (0, 0x303030303433);
-INSERT INTO `watchlist` VALUES (0, 0x303030303434);
-INSERT INTO `watchlist` VALUES (0, 0x303030303435);
-INSERT INTO `watchlist` VALUES (0, 0x303030303436);
-INSERT INTO `watchlist` VALUES (0, 0x303030303437);
-INSERT INTO `watchlist` VALUES (0, 0x303030303438);
-INSERT INTO `watchlist` VALUES (0, 0x303030303439);
-INSERT INTO `watchlist` VALUES (0, 0x303030303530);
-INSERT INTO `watchlist` VALUES (0, 0x303030303531);
-INSERT INTO `watchlist` VALUES (0, 0x303030303532);
-INSERT INTO `watchlist` VALUES (0, 0x303030303533);
-INSERT INTO `watchlist` VALUES (0, 0x303030303534);
-INSERT INTO `watchlist` VALUES (0, 0x303030303535);
-INSERT INTO `watchlist` VALUES (0, 0x303030303536);
-INSERT INTO `watchlist` VALUES (0, 0x303030303537);
-INSERT INTO `watchlist` VALUES (0, 0x303030303538);
-INSERT INTO `watchlist` VALUES (0, 0x303030303539);
-INSERT INTO `watchlist` VALUES (0, 0x303030303630);
-INSERT INTO `watchlist` VALUES (0, 0x303030303631);
-INSERT INTO `watchlist` VALUES (0, 0x303030303632);
-INSERT INTO `watchlist` VALUES (0, 0x303030303633);
-INSERT INTO `watchlist` VALUES (0, 0x303030303634);
-INSERT INTO `watchlist` VALUES (0, 0x303030303635);
-INSERT INTO `watchlist` VALUES (0, 0x303030303636);
-INSERT INTO `watchlist` VALUES (0, 0x303030303637);
-INSERT INTO `watchlist` VALUES (0, 0x303030303638);
-INSERT INTO `watchlist` VALUES (0, 0x303030303639);
-INSERT INTO `watchlist` VALUES (0, 0x303030303730);
-INSERT INTO `watchlist` VALUES (0, 0x303030303731);
-INSERT INTO `watchlist` VALUES (0, 0x303030303732);
-INSERT INTO `watchlist` VALUES (0, 0x303030303733);
-INSERT INTO `watchlist` VALUES (0, 0x303030303734);
-INSERT INTO `watchlist` VALUES (0, 0x303030303735);
-INSERT INTO `watchlist` VALUES (0, 0x303030303736);
-INSERT INTO `watchlist` VALUES (0, 0x303030303737);
-INSERT INTO `watchlist` VALUES (0, 0x303030303738);
-INSERT INTO `watchlist` VALUES (0, 0x303030303739);
-INSERT INTO `watchlist` VALUES (0, 0x303030303830);
-INSERT INTO `watchlist` VALUES (0, 0x303030303831);
-INSERT INTO `watchlist` VALUES (0, 0x303030303832);
-INSERT INTO `watchlist` VALUES (0, 0x303030303833);
-INSERT INTO `watchlist` VALUES (0, 0x303030303834);
-INSERT INTO `watchlist` VALUES (0, 0x303030303835);
-INSERT INTO `watchlist` VALUES (0, 0x303030303836);
-INSERT INTO `watchlist` VALUES (0, 0x303030303837);
-INSERT INTO `watchlist` VALUES (0, 0x303030303838);
-INSERT INTO `watchlist` VALUES (0, 0x303030303839);
-INSERT INTO `watchlist` VALUES (0, 0x303030303930);
-INSERT INTO `watchlist` VALUES (0, 0x303030303931);
-INSERT INTO `watchlist` VALUES (0, 0x303030303932);
-INSERT INTO `watchlist` VALUES (0, 0x303030303933);
-INSERT INTO `watchlist` VALUES (0, 0x303030303934);
-INSERT INTO `watchlist` VALUES (0, 0x303030303935);
-INSERT INTO `watchlist` VALUES (0, 0x303030303936);
-INSERT INTO `watchlist` VALUES (0, 0x303030303937);
-INSERT INTO `watchlist` VALUES (0, 0x303030303938);
-INSERT INTO `watchlist` VALUES (0, 0x303030303939);
-INSERT INTO `watchlist` VALUES (0, 0x30303030313030);
-INSERT INTO `watchlist` VALUES (1, 0x303030303031);
-INSERT INTO `watchlist` VALUES (1, 0x303030303033);
-
 SET FOREIGN_KEY_CHECKS = 1;
