+++ sfx-analyze/src/main/resources/application-prod.yaml
@@ -8,5 +8,5 @@ spring:
   cloud:
     nacos:
       discovery:
-        ip: 124.222.191.199 # 这里写服务器本身的地址
-      server-addr: http://150.158.161.165:8848 # Nacos地址
+        ip: localhost # 这里写服务器本身的地址
+      server-addr: localhost:8848 # Nacos地址
