+++ sfx-core/src/main/java/org/sfx/core/CoreApplication.java
@@ -3,11 +3,8 @@ package org.sfx.core;
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
-import org.springframework.context.annotation.PropertySource;
 
 @SpringBootApplication
-@PropertySource(value = {"classpath:sfx.properties"})
 @MapperScan(value = {"org.sfx.core.mapper"})
 public class CoreApplication {
