+++ sfx-gateway/src/main/resources/application.yml
@@ -27,6 +27,10 @@ Spring:
           uri: lb://analyzeservice
           predicates:
             - Path=/analyze/**
+        - id: stock-service
+          uri: lb://stockservice
+          predicates:
+            - Path=/stock/**
 server:
   port: 9000
 spring:
