+++ sfx-search/pom.xml
@@ -9,7 +9,7 @@
     </parent>
     <groupId>org.sfx.search</groupId>
     <artifactId>sfx-search</artifactId>
-    <version>1.0.0</version>
+    <version>1.0.1</version>
     <name>sfx-search</name>
     <description>sfx-backend-search</description>
     <properties>
