+++ sfx-search/src/main/java/org/sfx/search/controller/IndexController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@RequestMapping("index")
+@RequestMapping("/_search")
 @RestController
 public class IndexController {
