+++ sfx-search/src/main/resources/application-prod.yaml
@@ -2,5 +2,6 @@ spring:
   cloud:
     nacos:
       discovery:
-        ip: 124.222.191.199 # 这里写服务器本身的地址
-      server-addr: http://150.158.161.165:8848 # Nacos地址
+        ip: localhost # 这里写服务器本身的地址
+        port: 9002
+      server-addr: localhost:8848 # Nacos地址
